+++ app/client/cypress/e2e/Regression/ClientSide/IDE/Canvas_Context_Property_Pane_1_spec.js
@@ -180,147 +180,6 @@ describe("Canvas context Property Pane", function () {
     cy.get(propPaneBack).click();
     cy.get(".t--property-pane-title").should("contain", "Table1");
   });
-
-  it("10. Layered PropertyPane - Property Sections should retain state while switching between widgets, pages and Editor Panes", () => {
-    let propertySectionState = {
-      data: false,
-      general: true,
-    };
-
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("step");
-        setPropertyPaneSectionState(propertySectionState);
-      },
-      () => {
-        cy.wait(500);
-        verifyPropertyPaneSectionState(propertySectionState);
-      },
-      "Table1",
-    );
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
-
-  it("11. Layered PropertyPane - Property Tabs and Sections should retain state while switching between widgets, pages and Editor Panes", () => {
-    propertySectionState = {
-      textformatting: true,
-      color: false,
-    };
-
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("step");
-        cy.get(`.ads-v2-tabs__list-tab:contains("Style")`).eq(0).click();
-        setPropertyPaneSectionState(propertySectionState);
-      },
-      () => {
-        verifyPropertyPaneSectionState(propertySectionState);
-      },
-      "Table1",
-    );
-
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
-
-  it("12. Multi Layered PropertyPane - Code Editor should have focus while switching between widgets, pages and Editor Panes", function () {
-    let propertyControlSelector = ".t--property-control-text";
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("status");
-        cy.editColumn("menuIteme63irwbvnd", false);
-        cy.focusCodeInput(propertyControlSelector);
-      },
-      () => {
-        cy.assertSoftFocusOnCodeInput(propertyControlSelector);
-      },
-      "Table1",
-    );
-
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "status");
-
-    cy.wait(500);
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
-
-  it("13. Multi Layered PropertyPane - Toggle Property controls should have focus while switching between widgets, pages and Editor Panes", () => {
-    propertyControlSelector = `.t--property-control-visible input[type="checkbox"]`;
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("status");
-        cy.editColumn("menuIteme63irwbvnd", false);
-        cy.get(propertyControlSelector).click({ force: true });
-      },
-      () => {
-        cy.get(propertyControlSelector).should("be.focused");
-      },
-      "Table1",
-    );
-
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "status");
-
-    cy.wait(500);
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
-
-  it("14. Multi Layered PropertyPane - Property Sections should retain state while switching between widgets, pages and Editor Panes", () => {
-    let propertySectionState = {
-      basic: false,
-      general: true,
-    };
-
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("status");
-        cy.editColumn("menuIteme63irwbvnd", false);
-        setPropertyPaneSectionState(propertySectionState);
-      },
-      () => {
-        cy.wait(500);
-        verifyPropertyPaneSectionState(propertySectionState);
-      },
-      "Table1",
-    );
-
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "status");
-
-    cy.wait(500);
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
-
-  it("15. Multi Layered PropertyPane - Property Tabs and Sections should retain state while switching between widgets, pages and Editor Panes", () => {
-    propertySectionState = {
-      icon: true,
-      color: false,
-    };
-
-    verifyPropertyPaneContext(
-      () => {
-        cy.editColumn("status");
-        cy.editColumn("menuIteme63irwbvnd", false);
-        cy.get(`.ads-v2-tabs__list-tab:contains("Style")`).eq(0).click();
-        setPropertyPaneSectionState(propertySectionState);
-      },
-      () => {
-        verifyPropertyPaneSectionState(propertySectionState);
-      },
-      "Table1",
-    );
-
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "status");
-
-    cy.wait(500);
-    cy.get(propPaneBack).click();
-    cy.get(".t--property-pane-title").should("contain", "Table1");
-  });
 });
 
 let propertySectionClass = (section) =>
