+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/Chart/Chart_1_spec.js
@@ -233,59 +233,7 @@ describe("Chart Widget Functionality", function () {
     _.deployMode.DeployApp();
   });
 
-  it("12. Chart - Modal", function () {
-    //creating the Modal and verify Modal name
-    cy.createModal(this.dataSet.ModalName, "onDataPointClick");
-    _.deployMode.DeployApp();
-    cy.get(widgetsPage.chartPlotGroup).children().first().click();
-    cy.get(modalWidgetPage.modelTextField).should(
-      "have.text",
-      this.dataSet.ModalName,
-    );
-  });
-
-  it("13. Chart-Unckeck Visible field Validation", function () {
-    // Making the widget invisible
-    cy.togglebarDisable(commonlocators.visibleCheckbox);
-    _.deployMode.DeployApp();
-    cy.get(publish.chartWidget).should("not.exist");
-  });
-
-  it("14. Chart-Check Visible field Validation", function () {
-    // Making the widget visible
-    cy.togglebar(commonlocators.visibleCheckbox);
-    _.deployMode.DeployApp();
-    cy.get(publish.chartWidget).should("be.visible");
-  });
-
-  it("15. Toggle JS - Chart-Unckeck Visible field Validation", function () {
-    //Uncheck the disabled checkbox using JS and validate
-    cy.get(widgetsPage.toggleVisible).click({ force: true });
-    cy.testJsontext("visible", "false");
-    _.deployMode.DeployApp();
-    cy.get(publish.chartWidget).should("not.exist");
-  });
-
-  it("16. Toggle JS - Chart-Check Visible field Validation", function () {
-    //Check the disabled checkbox using JS and Validate
-    cy.testJsontext("visible", "true");
-    _.deployMode.DeployApp();
-    cy.get(publish.chartWidget).should("be.visible");
-  });
-
-  it("17. Chart Widget Functionality To Uncheck Horizontal Scroll Visible", function () {
-    cy.togglebarDisable(commonlocators.allowScroll);
-    _.deployMode.DeployApp();
-    cy.get(publish.horizontalTab).should("not.exist");
-  });
-
-  it("18. Chart Widget Functionality To Check Horizontal Scroll Visible", function () {
-    cy.togglebar(commonlocators.allowScroll);
-    _.deployMode.DeployApp();
-    cy.get(publish.horizontalTab).eq(1).should("exist");
-  });
-
-  it("19. Check Chart widget reskinning config", function () {
+  it("12. Check Chart widget reskinning config", function () {
     cy.get(widgetsPage.toggleChartType).click({ force: true });
     cy.UpdateChartType("Column chart");
