+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/JSONForm/JSONForm_HiddenFields_1_spec.js
@@ -164,78 +164,4 @@ describe("JSON Form Hidden fields", () => {
       city: "Bangalore",
     });
   });
-
-  it("9. can hide Phone Number Input Field", () => {
-    const defaultValue = "1000";
-    // Add new custom field
-    addCustomField("Phone Number Input");
-
-    cy.testJsontext("defaultvalue", defaultValue);
-
-    hideAndVerifyProperties("customField1", defaultValue);
-
-    removeCustomField();
-  });
-
-  it("10. can hide Radio Group Field", () => {
-    const defaultValue = "Y";
-    // Add new custom field
-    addCustomField("Phone Number Input");
-
-    cy.testJsontext("defaultvalue", defaultValue);
-
-    hideAndVerifyProperties("customField1", defaultValue);
-
-    removeCustomField();
-  });
-
-  it("11. can hide Select Field", () => {
-    const defaultValue = "BLUE";
-    // Add new custom field
-    addCustomField(/^Select/);
-
-    cy.testJsontext("defaultselectedvalue", defaultValue);
-
-    hideAndVerifyProperties("customField1", defaultValue);
-
-    removeCustomField();
-  });
-
-  it("12. can hide Switch Field", () => {
-    // Add new custom field
-    addCustomField("Switch");
-
-    hideAndVerifyProperties("customField1", true);
-
-    removeCustomField();
-  });
-
-  it("13. hides fields on first load", () => {
-    cy.openPropertyPane("jsonformwidget");
-
-    // hide education field
-    cy.openFieldConfiguration("education");
-    cy.togglebarDisable(".t--property-control-visible input");
-    cy.get(backBtn).click({ force: true }).wait(500);
-    // hide name field
-    cy.openFieldConfiguration("name");
-    cy.togglebarDisable(".t--property-control-visible input");
-
-    // publish the app
-    deployMode.DeployApp();
-    cy.wait(4000);
-
-    // Check if name is hidden
-    cy.get(`${fieldPrefix}-name`).should("not.exist");
-    // Check if education is hidden
-    cy.get(`${fieldPrefix}-education`).should("not.exist");
-
-    // check if name and education are not present in form data
-    cy.get(".t--widget-textwidget .bp3-ui-text").then(($el) => {
-      const formData = JSON.parse($el.text());
-
-      cy.wrap(formData.name).should("deep.equal", undefined);
-      cy.wrap(formData.education).should("deep.equal", undefined);
-    });
-  });
 });
