+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/List/List4_1_spec.js
@@ -210,57 +210,4 @@ describe("Container Widget Functionality", function () {
     );
     _.deployMode.NavigateBacktoEditor();
   });
-
-  it("11. Add new item in the list widget array object", function () {
-    // Open Property pane
-    _.entityExplorer.SelectEntityByName("List1", "Widgets");
-
-    //Add the new item in the list
-    _.propPane.UpdatePropertyFieldValue(
-      "Items",
-      JSON.stringify(this.dataSet.ListItems),
-    );
-    cy.wait(2000);
-    _.deployMode.DeployApp();
-    _.deployMode.NavigateBacktoEditor();
-  });
-
-  it("12. Adding large item Spacing for item card", function () {
-    // Open Property pane
-    _.entityExplorer.SelectEntityByName("List1", "Widgets");
-    _.propPane.MoveToTab("Style");
-    // Scroll down to Styles and Add item spacing for item card
-    cy.testJsontext("itemspacing\\(" + "px" + "\\)", 12);
-    cy.wait(2000);
-    // Click on Deploy and ensure it is deployed appropriately
-    _.deployMode.DeployApp();
-    _.deployMode.NavigateBacktoEditor();
-  });
-
-  it("13. Renaming the widget from Property pane and Entity explorer ", function () {
-    // Open Property pane
-    _.entityExplorer.SelectEntityByName("List1", "Widgets");
-
-    // Change the list widget name from property pane and Verify it
-    cy.widgetText(
-      "List2",
-      widgetsPage.listWidgetName,
-      widgetsPage.widgetNameSpan,
-    );
-    // Change the list widget name from Entity Explorer
-    cy.renameEntity("List2", "List1");
-    // Mouse over to list name
-    _.entityExplorer.SelectEntityByName("List1");
-
-    cy.get(widgetsPage.listWidgetName)
-      .first()
-      .trigger("mouseover", { force: true });
-    // Verify the list name is changed
-    cy.contains(
-      widgetsPage.listWidgetName + " " + commonlocators.listWidgetNameTag,
-      "List1",
-    );
-    _.deployMode.DeployApp();
-    _.deployMode.NavigateBacktoEditor();
-  });
 });
