+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/TableV1/TableFilter1_2_Spec.ts
@@ -7,7 +7,7 @@ describe("Verify various Table_Filter combinations", function () {
     });
   });
 
-  it("1. Adding Data to Table Widget", function () {
+  it("1. Verify Table Filter for 'empty'", function () {
     _.entityExplorer.SelectEntityByName("Table1");
     _.propPane.UpdatePropertyFieldValue(
       "Table data",
@@ -16,113 +16,13 @@ describe("Verify various Table_Filter combinations", function () {
     _.assertHelper.AssertNetworkStatus("@updateLayout", 200);
     _.agHelper.PressEscape();
     _.deployMode.DeployApp();
-  });
 
-  it("2. Table Widget Search Functionality", function () {
-    _.table.ReadTableRowColumnData(1, 3, "v1", 2000).then((cellData) => {
-      expect(cellData).to.eq("Lindsay Ferguson");
-      _.table.SearchTable(cellData);
-      _.table.ReadTableRowColumnData(0, 3).then((afterSearch) => {
-        expect(afterSearch).to.eq("Lindsay Ferguson");
-      });
-    });
-    _.table.RemoveSearchTextNVerify("2381224");
-
-    _.table.SearchTable("7434532");
-    _.table.ReadTableRowColumnData(0, 3).then((afterSearch) => {
-      expect(afterSearch).to.eq("Byron Fields");
-    });
-    _.table.RemoveSearchTextNVerify("2381224");
-  });
-
-  it("3. Verify Table Filter for 'contain'", function () {
-    _.table.OpenNFilterTable("userName", "contains", "Lindsay");
-    _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
-      expect($cellData).to.eq("Lindsay Ferguson");
-    });
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("4. Verify Table Filter for 'does not contain'", function () {
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Tuna Salad");
-    });
-    _.table.OpenNFilterTable("productName", "does not contain", "Tuna");
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Beef steak");
-    });
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("5. Verify Table Filter for 'starts with'", function () {
-    _.table.ReadTableRowColumnData(4, 4).then(($cellData) => {
-      expect($cellData).to.eq("Avocado Panini");
-    });
-    _.table.OpenNFilterTable("productName", "starts with", "Avo");
-    _.table.ReadTableRowColumnData(0, 4).then(($cellData) => {
-      expect($cellData).to.eq("Avocado Panini");
-    });
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("6. Verify Table Filter for 'ends with' - case sensitive", function () {
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Tuna Salad");
-    });
-    _.table.OpenNFilterTable("productName", "ends with", "wich");
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Chicken Sandwich");
-    });
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("7. Verify Table Filter for 'ends with' - case insenstive", function () {
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Tuna Salad");
-    });
-    _.table.OpenNFilterTable("productName", "ends with", "WICH");
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Chicken Sandwich");
-    });
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("8. Verify Table Filter for 'ends with' - on wrong column", function () {
-    _.table.ReadTableRowColumnData(1, 4).then(($cellData) => {
-      expect($cellData).to.eq("Tuna Salad");
-    });
-    _.table.OpenNFilterTable("userName", "ends with", "WICH");
-    _.table.WaitForTableEmpty();
-    _.table.RemoveFilterNVerify("2381224");
-  });
-
-  it("9. Verify Table Filter for 'is exactly' - case sensitive", function () {
-    _.table.ReadTableRowColumnData(2, 4).then(($cellData) => {
-      expect($cellData).to.eq("Beef steak");
-    });
-    _.table.OpenNFilterTable("productName", "is exactly", "Beef steak");
-    _.table.ReadTableRowColumnData(0, 4).then(($cellData) => {
-      expect($cellData).to.eq("Beef steak");
-    });
-    _.table.RemoveFilterNVerify("2381224", true);
-  });
-
-  it("10. Verify Table Filter for 'is exactly' - case insensitive", function () {
-    _.table.ReadTableRowColumnData(2, 4).then(($cellData) => {
-      expect($cellData).to.eq("Beef steak");
-    });
-    _.table.OpenNFilterTable("productName", "is exactly", "Beef STEAK");
-    _.table.WaitForTableEmpty();
-    _.table.RemoveFilterNVerify("2381224", true);
-  });
-
-  it("11. Verify Table Filter for 'empty'", function () {
     _.table.OpenNFilterTable("email", "empty");
     _.table.WaitForTableEmpty();
     _.table.RemoveFilterNVerify("2381224");
   });
 
-  it("12. Verify Table Filter for 'not empty'", function () {
+  it("2. Verify Table Filter for 'not empty'", function () {
     _.table.ReadTableRowColumnData(4, 5).then(($cellData) => {
       expect($cellData).to.eq("7.99");
     });
@@ -133,7 +33,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224");
   });
 
-  it("13. Verify Table Filter - Where Edit - Change condition along with input value", function () {
+  it("3. Verify Table Filter - Where Edit - Change condition along with input value", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
@@ -160,7 +60,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("14. Verify Table Filter - Where Edit - Single Column, Condition & input value", function () {
+  it("4. Verify Table Filter - Where Edit - Single Column, Condition & input value", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
@@ -209,7 +109,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("15. Verify Table Filter for OR operator - different row match", function () {
+  it("5. Verify Table Filter for OR operator - different row match", function () {
     _.table.ReadTableRowColumnData(2, 3).then(($cellData) => {
       expect($cellData).to.eq("Tobias Funke");
     });
@@ -225,7 +125,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("16. Verify Table Filter for OR operator - same row match", function () {
+  it("6. Verify Table Filter for OR operator - same row match", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
@@ -240,7 +140,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("17. Verify Table Filter for OR operator - two 'ORs'", function () {
+  it("7. Verify Table Filter for OR operator - two 'ORs'", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
@@ -259,7 +159,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("18. Verify Table Filter for AND operator - different row match", function () {
+  it("8. Verify Table Filter for AND operator - different row match", function () {
     _.table.ReadTableRowColumnData(3, 3).then(($cellData) => {
       expect($cellData).to.eq("Byron Fields");
     });
@@ -278,7 +178,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("19. Verify Table Filter for AND operator - same row match", function () {
+  it("9. Verify Table Filter for AND operator - same row match", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
@@ -293,7 +193,7 @@ describe("Verify various Table_Filter combinations", function () {
     _.table.RemoveFilterNVerify("2381224", true, false);
   });
 
-  it("20. Verify Table Filter for AND operator - same row match - edit input text value", function () {
+  it("10. Verify Table Filter for AND operator - same row match - edit input text value", function () {
     _.table.ReadTableRowColumnData(0, 3).then(($cellData) => {
       expect($cellData).to.eq("Michael Lawson");
     });
